プラグインをデフォルトモジュールにできるようにする。 選択したプラグインが存在しないことも考慮する
