+++ WindowTranslator.Abstractions/ComponentModel/DefaultModuleAttribute.cs
@@ -1,0 +1,7 @@
+﻿namespace WindowTranslator.ComponentModel;
+
+/// <summary>
+/// デフォルトで使用するモジュールを示す属性
+/// </summary>
+[AttributeUsage(AttributeTargets.Class)]
+public class DefaultModuleAttribute : Attribute;
