--- WindowTranslator/ComponentModel/DefaultModuleAttribute.cs
@@ -1,6 +1,0 @@
-﻿namespace WindowTranslator.ComponentModel;
-
-[AttributeUsage(AttributeTargets.Class)]
-internal class DefaultModuleAttribute : Attribute
-{
-}
