+++ WindowTranslator/appsettings.json
@@ -1,8 +1,5 @@
 {
   "ViewMode": 1,
-  "SelectedPlugins": {
-    "ITranslateModule": "DeepLTranslator"
-  },
   "Sentry": {
     "Dsn": "https://5301a4e1d7cd4c45877493fd27303166@o351180.ingest.sentry.io/6471678"
   }
